`edit` now takes `Element` rather than `String elementId`
ace supports both, but internally they just do a
`document.getElementById`, so there is little point to it, and that is
also a design that was made in a pre-shadow-dom world.  easier to leave
the element resolution to the user, esp. since we cannot have multiple
overloads for the `edit` method