Selection : doc and test for `moveCursorWordLeft` and `moveCursorWordRight`
also removed the equivalent overloads for 'long' and 'short' words.
From inspecting the javascript all of these methods lead to the same
code path because the variable `$selectLongWords` does not exist.  They
probably have kept the overloads to not break anyone?  But since the
overloads are just confusing and mean nothing anymore, it is best not to
expose them.