_HasProxy : make `_onHasProxy` and `_hasProxy` private as these are an implementation concern
This should not be exposed to the user, as they should not need to know
these concerns for the most part.  The `dispose` method is still public,
for releasing the proxy.  Over time the implementation may change to
pure dart and the `hasProxy` would no longer make sense.  A `dispose`
might always make sense, because it would still close the event streams.
In the least it could remain as a noop.